move EventSourcing.Persistence DI code into main EventSourcing.Persistence package
