--- EventSourcing.Persistence/Extensions.DependencyInjection.Persistence/Extensions.DependencyInjection.Persistence.csproj
@@ -1,34 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <TreatWarningsAsErrors Condition=" '$(Configuration)' == 'Release' ">true</TreatWarningsAsErrors>
-        <GenerateDocumentationFile>true</GenerateDocumentationFile>
-        <RootNamespace>EventSourcing.Extensions.DependencyInjection.Persistence</RootNamespace>
-        <AssemblyName>EventSourcing.Extensions.DependencyInjection.Persistence</AssemblyName>
-    </PropertyGroup>
-
-    <!-- NuGet metadata -->
-    <PropertyGroup>
-        <Title>EventSourcing.Extensions.DependencyInjection.Persistence</Title>
-        <Description>This package contains DependencyInjection extensions for default implementation of EventSourcing.Persistence.Abstraction package.</Description>
-        <Authors>Dawid Szmidka</Authors>
-        <PackageId>EventSourcing.Extensions.DependencyInjection.Persistence</PackageId>
-        <Version>0.1.0-alpha</Version>
-        <PackageReleaseNotes>
-            - Initial version of EventSourcing.Extensions.DependencyInjection.Persistence project
-        </PackageReleaseNotes>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <PackageTags>EventSourcing;EventSourcing.Extensions;EventSourcing.Extensions.DependencyInjection</PackageTags>
-        <IncludeSymbols>true</IncludeSymbols>
-        <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-        <IncludeSource>true</IncludeSource>
-        <RepositoryUrl>https://github.com/TheName/EventSourcing</RepositoryUrl>
-    </PropertyGroup>
-    
-    <ItemGroup>
-      <ProjectReference Include="..\..\Core\EventSourcing\EventSourcing.csproj" />
-      <ProjectReference Include="..\Persistence\Persistence.csproj" />
-    </ItemGroup>
-
-</Project>
