+++ EventSourcing.Persistence/PostgreSql/Persistence.PostgreSql/Persistence.PostgreSql.csproj
@@ -27,11 +27,11 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\Extensions.DependencyInjection.Persistence\Extensions.DependencyInjection.Persistence.csproj" />
         <ProjectReference Include="..\..\Persistence\Persistence.csproj" />
     </ItemGroup>
 
     <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="5.0.0" />
       <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />
       <PackageReference Include="Npgsql" Version="6.0.3" />
     </ItemGroup>
