+++ EventSourcing.Persistence/SqlServer/Persistence.SqlServer/Extensions/EventSourcingBuilderExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Data.SqlClient;
 using EventSourcing.Abstractions.DependencyInjection;
-using EventSourcing.Extensions.DependencyInjection.Persistence;
+using EventSourcing.Persistence.Extensions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
