+++ EventSourcing.Persistence/SqlServer/Persistence.SqlServer/Persistence.SqlServer.csproj
@@ -27,12 +27,12 @@
     </PropertyGroup>
     
     <ItemGroup>
-      <ProjectReference Include="..\..\Extensions.DependencyInjection.Persistence\Extensions.DependencyInjection.Persistence.csproj" />
       <ProjectReference Include="..\..\Persistence\Persistence.csproj" />
     </ItemGroup>
     
     <ItemGroup>
       <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />
+      <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="5.0.0" />
       <PackageReference Include="System.Data.SqlClient" Version="4.8.2" />
     </ItemGroup>
