+++ EventSourcing.sln
@@ -16,8 +16,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extensions.DatabaseMigratio
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extensions.DatabaseMigrations.Persistence.SqlServer.DbUp", "EventSourcing.Persistence\SqlServer\Extensions.DatabaseMigrations.Persistence.SqlServer.DbUp\Extensions.DatabaseMigrations.Persistence.SqlServer.DbUp.csproj", "{6A449FAA-DB82-4988-B5BF-B4BAD8C4C9E2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extensions.DependencyInjection.Persistence", "EventSourcing.Persistence\Extensions.DependencyInjection.Persistence\Extensions.DependencyInjection.Persistence.csproj", "{9F569390-1B76-4D6C-98FC-87D2F7F6C051}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence", "EventSourcing.Persistence\Persistence\Persistence.csproj", "{44E808CF-ABC2-4903-B9A8-A25F25079643}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.Abstractions", "EventSourcing.Persistence\Persistence.Abstractions\Persistence.Abstractions.csproj", "{A50EA1B7-9778-4AEB-88B4-3F9563646F99}"
@@ -91,8 +89,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.PostgreSql.IntegrationTests", "EventSourcing.Persistence\PostgreSql\Persistence.PostgreSql.IntegrationTests\Persistence.PostgreSql.IntegrationTests.csproj", "{246BB686-60C1-45E2-B435-719C55015E6F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.IntegrationTests.Base", "EventSourcing.Persistence\Persistence.IntegrationTests.Base\Persistence.IntegrationTests.Base.csproj", "{DC12019B-B5C2-4AB1-A0B1-89372FC05C48}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extensions.DependencyInjection.Persistence.UnitTests", "EventSourcing.Persistence\Extensions.DependencyInjection.Persistence.UnitTests\Extensions.DependencyInjection.Persistence.UnitTests.csproj", "{28083AC4-9515-45C8-8AD1-2FD780FBBDF0}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extensions.DependencyInjection.Serialization.UnitTests", "EventSourcing.Serialization\Extensions.DependencyInjection.Serialization.UnitTests\Extensions.DependencyInjection.Serialization.UnitTests.csproj", "{C92098C2-9DEB-466E-9AFB-D0F6700CDD85}"
 EndProject
@@ -172,10 +168,6 @@ Global
 		{6A449FAA-DB82-4988-B5BF-B4BAD8C4C9E2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6A449FAA-DB82-4988-B5BF-B4BAD8C4C9E2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6A449FAA-DB82-4988-B5BF-B4BAD8C4C9E2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9F569390-1B76-4D6C-98FC-87D2F7F6C051}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9F569390-1B76-4D6C-98FC-87D2F7F6C051}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9F569390-1B76-4D6C-98FC-87D2F7F6C051}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9F569390-1B76-4D6C-98FC-87D2F7F6C051}.Release|Any CPU.Build.0 = Release|Any CPU
 		{44E808CF-ABC2-4903-B9A8-A25F25079643}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{44E808CF-ABC2-4903-B9A8-A25F25079643}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{44E808CF-ABC2-4903-B9A8-A25F25079643}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -300,10 +292,6 @@ Global
 		{DC12019B-B5C2-4AB1-A0B1-89372FC05C48}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DC12019B-B5C2-4AB1-A0B1-89372FC05C48}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DC12019B-B5C2-4AB1-A0B1-89372FC05C48}.Release|Any CPU.Build.0 = Release|Any CPU
-		{28083AC4-9515-45C8-8AD1-2FD780FBBDF0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{28083AC4-9515-45C8-8AD1-2FD780FBBDF0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{28083AC4-9515-45C8-8AD1-2FD780FBBDF0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{28083AC4-9515-45C8-8AD1-2FD780FBBDF0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C92098C2-9DEB-466E-9AFB-D0F6700CDD85}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C92098C2-9DEB-466E-9AFB-D0F6700CDD85}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C92098C2-9DEB-466E-9AFB-D0F6700CDD85}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -390,7 +378,6 @@ Global
 		{90AA64F3-52A1-45E0-AD2E-62ACA6C276F8} = {3829A2ED-C973-4C40-89B2-276296AC020A}
 		{A2CE13CB-9C3F-49A2-8109-BE1C02BAA25D} = {3829A2ED-C973-4C40-89B2-276296AC020A}
 		{7B67F03F-428F-40B2-86F9-B5ED8E3EC1E3} = {3829A2ED-C973-4C40-89B2-276296AC020A}
-		{9F569390-1B76-4D6C-98FC-87D2F7F6C051} = {8792BDC7-C64B-4270-A84C-579204AC79B5}
 		{44E808CF-ABC2-4903-B9A8-A25F25079643} = {8792BDC7-C64B-4270-A84C-579204AC79B5}
 		{A50EA1B7-9778-4AEB-88B4-3F9563646F99} = {8792BDC7-C64B-4270-A84C-579204AC79B5}
 		{F5BF6C9E-4709-4E98-810F-9F3F8A87D021} = {8792BDC7-C64B-4270-A84C-579204AC79B5}
@@ -426,7 +413,6 @@ Global
 		{B9719B28-3265-4232-AB7E-04BB1CA55DA1} = {8B8FC55F-39CF-431D-AF1B-D67A24CB4071}
 		{246BB686-60C1-45E2-B435-719C55015E6F} = {8B8FC55F-39CF-431D-AF1B-D67A24CB4071}
 		{DC12019B-B5C2-4AB1-A0B1-89372FC05C48} = {8792BDC7-C64B-4270-A84C-579204AC79B5}
-		{28083AC4-9515-45C8-8AD1-2FD780FBBDF0} = {8792BDC7-C64B-4270-A84C-579204AC79B5}
 		{C92098C2-9DEB-466E-9AFB-D0F6700CDD85} = {29745774-1ACA-4563-89CD-8C145BAF9D29}
 		{F65145F4-5FE1-4A39-8710-9A1417BA5001} = {29745774-1ACA-4563-89CD-8C145BAF9D29}
 		{6520A5D9-FD64-48E8-95B4-A047D543BBF1} = {29745774-1ACA-4563-89CD-8C145BAF9D29}
